+++ src/components/modals/modal.utils.js
@@ -1,11 +1,11 @@
-import { generateModalFragment } from "../modals/modal.manager.js";
-import { addNewClassModal } from "../modals/templates-modal/add.class.modal.js";
-import { addNewTeacherModal } from "../modals/templates-modal/add.teacher.modal.js";
-import { addNewStudentModal } from "../modals/templates-modal/add.student.modal.js";
+import { generateModalFragment } from "./modal.js";
+import { addNewClassModal } from "../modals/add-modals/add.class.modal.js";
+import { addNewTeacherModal } from "../modals/add-modals/add.teacher.modal.js";
+import { addNewStudentModal } from "../modals/add-modals/add.student.modal.js";
 import { manageSavingEvents } from "../../functions.js";
 import { showAddingModal } from "../../utils.js";
 
-export function generateNewObjectModal() {
+export function generateNewModal() {
   const pEl = document.getElementById("add-text");
   const action = pEl.getAttribute("data-action");
   switch (action) {
