+++ src/components/school/school.js
@@ -20,15 +20,15 @@ let storedSchoolData = JSON.parse(localStorage.getItem("school"));
 export const LMSchool = new School();
 if (storedSchoolData) {
   if (storedSchoolData.classes && storedSchoolData.classes.length > 0) {
-    LMSchool.classes = storedSchoolData.classes
+    LMSchool.classes = storedSchoolData.classes;
   }
 
   if (storedSchoolData.teachers && storedSchoolData.teachers.length > 0) {
-    LMSchool.teachers = storedSchoolData.teachers
+    LMSchool.teachers = storedSchoolData.teachers;
   }
 
   if (storedSchoolData.students && storedSchoolData.students.length > 0) {
-    LMSchool.students = storedSchoolData.students
+    LMSchool.students = storedSchoolData.students;
   }
 }
