+++ src/pages/students/components/student.card.generator.js
@@ -1,5 +1,5 @@
 import { students } from "../../../service/data.js";
-import { findAverageGradeOfStudent } from "../../../functions.js";
+import { findAverageGradeOfStudent } from "../../../components/student/student.utils.js";
 
 export function generateStudentCardsHTML() {
   let cardsHTML = "";
@@ -16,10 +16,14 @@ export function generateStudentCardsHTML() {
           <div class="card border-info" style="width: 18rem;">
             <div class="card-body">
             <div class="d-flex justify-content-end mb-2">
-            <button id="${student.id}" data-action="student" class="edit-button btn btn-primary btn-sm">
+            <button id="${
+              student.id
+            }" data-action="student" class="edit-button btn btn-primary btn-sm">
             Edit
             </button>
-            <button id="${student.id}" data-action="student" class="remove-button btn btn-danger btn-sm mx-2">
+            <button id="${
+              student.id
+            }" data-action="student" class="remove-button btn btn-danger btn-sm mx-2">
             Remove
             </button>
               </div>
@@ -44,5 +48,3 @@ export function generateStudentCardsHTML() {
 
   return cardsHTML;
 }
-
-
